multiple items, see full commit message, closes #5
I added the code changes needed to support jsonb to address the
deprecation warning from langchain.
I also reverted changes made to the database initialization in the
previous commit. I mistakenly did the wrong test to validate those
changes and found they did not actually work and I identified a better method.

I updated the code to delete any existing tables with {database_name}
per the reasons detailed in #5. I initially planned to re-initialize the
database, but that approach is out of line with how things should be
done in CF and was likely overkill anyway. I updated my approach to only
re-initialize the table which is a better approach.

To make this change I had to ensure the control flow followed a specific
order which was complicated because the code is disorganized and mixed
function creation with control flow, so I also reorginized the code more
logically for better readability and alignment with best practices.